README: Suggest '/param:url(*)' as param-based route path
This prevents the need to use double-slashes (which is needed when configured with '/param/:url')

Example:
  image: /foo.jpg
  expected (and nice) sharpie path: /param/foo.jpg
  currently required sharpie path:  /param//foo.jpg

  As sharpie uses URL.resolve to map to the original image path, these url's will map to:
    /param/foo.jpg  => /param/foo.jpg (infinte loop)
    /param//foo.jpg => /foo.jpg

That means the expected path creates an infinite loop as it will request the resource from sharpie again and again.
This patch fixes that behaviour by adding the leading slash to the `url` parameter

As a site effect :url(*) allows to use nested image paths e.g. /param/path/to/image.jpg